게시글 이미지 수정하기 완료, 팔로우 리스트 불러오기 에러 해결 완료 좋아요 실시간 반영 안됨 좋아요 axios요청 store로 이동 작업 시작 팔로우 리스트 마운트 시 불러오기 path 문제로 클릭 시에만 불러와 짐 수정 예정

+++ src/main/java/com/service/BoardServiceImpl.java
@@ -43,18 +43,20 @@ public class BoardServiceImpl implements BoardService {
 	public void boardUpdate(BoardDto boardDto, MultipartHttpServletRequest image) throws Exception {
 		
 		boardMapper.boardUpdate(boardDto);
+		
 		if (!"".equals(Integer.toString(boardDto.getBoardSeq())) & Integer.toString(boardDto.getBoardSeq()) != null) {
-			
 			List<BoardFileDto> list = fileutils.parseFileInfo(boardDto, image);
-			int boardImgLength = boardMapper.boardImgLengthCheck(boardDto.getBoardSeq());
-			int ReboardImgLength = list.size();
-			
-			if (boardImgLength == ReboardImgLength) {
-				boardMapper.boardImgUpdate(list);
-			}
-			else {
-				boardMapper.deleteBoardImg(boardDto.getBoardSeq());
-				boardMapper.boardImgInsert(list);
+			if (!list.isEmpty()) {
+				int boardImgLength = boardMapper.boardImgLengthCheck(boardDto.getBoardSeq());
+				int ReboardImgLength = list.size();
+				
+				if (boardImgLength == ReboardImgLength) {
+					boardMapper.boardImgUpdate(list);
+				}
+				else {
+					boardMapper.deleteBoardImg(boardDto.getBoardSeq());
+					boardMapper.boardImgInsert(list);
+				}
 			}
 		}
 	}
